+++ nsmd/nsmPort/nsmGpuPciePort.cpp
@@ -37,7 +37,8 @@ NsmGpuPciePort::NsmGpuPciePort(
     sdbusplus::bus::bus& bus, const std::string& name, const std::string& type,
     const std::string& health, const std::string& chasisState,
     const std::vector<utils::Association>& associations,
-    const std::string& inventoryObjPath) : NsmObject(name, type)
+    const std::string& inventoryObjPath) :
+    NsmObject(name, type)
 {
     lg2::info("NsmGpuPciePort: create sensor:{NAME}", "NAME", name.c_str());
     associationDefIntf =
@@ -66,7 +67,8 @@ NsmGpuPciePortInfo::NsmGpuPciePortInfo(
     const std::string& name, const std::string& type,
     const std::string& portType, const std::string& portProtocol,
     std::shared_ptr<PortInfoIntf> portInfoIntf) :
-    NsmObject(name, type), portInfoIntf(portInfoIntf)
+    NsmObject(name, type),
+    portInfoIntf(portInfoIntf)
 {
     lg2::info("NsmGpuPciePortInfo: create sensor:{NAME}", "NAME", name.c_str());
     portInfoIntf->type(PortInfoIntf::convertPortTypeFromString(portType));
@@ -77,8 +79,8 @@ NsmGpuPciePortInfo::NsmGpuPciePortInfo(
 NsmClearPCIeCounters::NsmClearPCIeCounters(
     const std::string& name, const std::string& type, const uint8_t groupId,
     uint8_t deviceIndex, std::shared_ptr<ClearPCIeIntf> clearPCIeIntf) :
-    NsmObject(name, type), groupId(groupId), deviceIndex(deviceIndex),
-    clearPCIeIntf(clearPCIeIntf)
+    NsmObject(name, type),
+    groupId(groupId), deviceIndex(deviceIndex), clearPCIeIntf(clearPCIeIntf)
 {
     lg2::info("NsmClearPCIeCounters:: create sensor {NAME} of group {GROUP_ID}",
               "NAME", name, "GROUP_ID", groupId);
@@ -421,18 +423,17 @@ static requester::Coroutine createNsmGpuPcieSensor(SensorManager& manager,
                 bus, name, type, health, chasisState, associations,
                 inventoryObjPath);
             nsmDevice->deviceSensors.emplace_back(sensor);
-            unsigned long long deviceIndex{};
+            uint64_t deviceIndex{};
             if (allBaseIfaceProperties.count("DeviceIndex"))
             {
-                deviceIndex = std::get<unsigned long long>(
+                deviceIndex = std::get<uint64_t>(
                     allBaseIfaceProperties.at("DeviceIndex"));
             }
-            std::vector<unsigned long long> clearableScalarGroup{};
+            std::vector<uint64_t> clearableScalarGroup{};
             if (allBaseIfaceProperties.count("ClearableScalarGroup"))
             {
-                clearableScalarGroup =
-                    std::get<std::vector<unsigned long long>>(
-                        allBaseIfaceProperties.at("ClearableScalarGroup"));
+                clearableScalarGroup = std::get<std::vector<uint64_t>>(
+                    allBaseIfaceProperties.at("ClearableScalarGroup"));
             }
 
             auto clearPCIeIntf = std::make_shared<NsmClearPCIeIntf>(
@@ -513,10 +514,10 @@ static requester::Coroutine createNsmGpuPcieSensor(SensorManager& manager,
                 priority =
                     std::get<bool>(allCurrentIfaceProperties.at("Priority"));
             }
-            unsigned long long deviceIndex{};
+            uint64_t deviceIndex{};
             if (allBaseIfaceProperties.count("DeviceIndex"))
             {
-                deviceIndex = std::get<unsigned long long>(
+                deviceIndex = std::get<uint64_t>(
                     allBaseIfaceProperties.at("DeviceIndex"));
             }
