+++ nsmd/nsmPort/nsmRetimerPort.cpp
@@ -16,7 +16,8 @@ NsmPort::NsmPort(sdbusplus::bus::bus& bus, std::string& portName,
                  const std::string& type,
                  const std::vector<utils::Association>& associations,
                  const std::string& inventoryObjPath) :
-    NsmObject(portName, type), portName(portName)
+    NsmObject(portName, type),
+    portName(portName)
 {
     lg2::info("NsmPCIePort: create sensor:{NAME}", "NAME", portName.c_str());
     portIntf = std::make_unique<PortIntf>(bus, inventoryObjPath.c_str());
@@ -41,8 +42,9 @@ NsmPCIeECCGroup1::NsmPCIeECCGroup1(const std::string& name,
                                    std::shared_ptr<PortInfoIntf> portInfoIntf,
                                    std::shared_ptr<PortWidthIntf> portWidthIntf,
                                    uint8_t deviceIndex) :
-    NsmPcieGroup(name, type, deviceIndex, GROUP_ID_1), objPath(inventoryPath),
-    portInfoIntf(portInfoIntf), portWidthIntf(portWidthIntf)
+    NsmPcieGroup(name, type, deviceIndex, GROUP_ID_1),
+    objPath(inventoryPath), portInfoIntf(portInfoIntf),
+    portWidthIntf(portWidthIntf)
 {
     lg2::info("NsmPCIeECCGroup1: {NAME}", "NAME", name.c_str());
 
@@ -174,8 +176,8 @@ NsmPCIeECCGroup2::NsmPCIeECCGroup2(const std::string& name,
                                    const std::string& inventoryPath,
                                    std::shared_ptr<PCIeEccIntf> pcieEccIntf,
                                    uint8_t deviceIndex) :
-    NsmPcieGroup(name, type, deviceIndex, GROUP_ID_2), objPath(inventoryPath),
-    pcieEccIntf(pcieEccIntf)
+    NsmPcieGroup(name, type, deviceIndex, GROUP_ID_2),
+    objPath(inventoryPath), pcieEccIntf(pcieEccIntf)
 {
     lg2::info("NsmPCIeECCGroup2: {NAME}", "NAME", name.c_str());
 
@@ -269,8 +271,8 @@ NsmPCIeECCGroup3::NsmPCIeECCGroup3(const std::string& name,
                                    const std::string& inventoryPath,
                                    std::shared_ptr<PCIeEccIntf> pcieEccIntf,
                                    uint8_t deviceIndex) :
-    NsmPcieGroup(name, type, deviceIndex, GROUP_ID_3), objPath(inventoryPath),
-    pcieEccIntf(pcieEccIntf)
+    NsmPcieGroup(name, type, deviceIndex, GROUP_ID_3),
+    objPath(inventoryPath), pcieEccIntf(pcieEccIntf)
 {
     lg2::info("NsmPCIeECCGroup3: {NAME}", "NAME", name.c_str());
 
@@ -337,8 +339,8 @@ NsmPCIeECCGroup4::NsmPCIeECCGroup4(const std::string& name,
                                    const std::string& inventoryPath,
                                    std::shared_ptr<PCIeEccIntf> pcieEccIntf,
                                    uint8_t deviceIndex) :
-    NsmPcieGroup(name, type, deviceIndex, GROUP_ID_4), objPath(inventoryPath),
-    pcieEccIntf(pcieEccIntf)
+    NsmPcieGroup(name, type, deviceIndex, GROUP_ID_4),
+    objPath(inventoryPath), pcieEccIntf(pcieEccIntf)
 {
     lg2::info("NsmPCIeECCGroup4: {NAME}", "NAME", name.c_str());
 
@@ -608,17 +610,16 @@ static requester::Coroutine
     {
         priority = std::get<bool>(allCurrentIfaceProperties.at("Priority"));
     }
-    unsigned long long count{};
+    uint64_t count{};
     if (allCurrentIfaceProperties.count("Count"))
     {
-        count =
-            std::get<unsigned long long>(allCurrentIfaceProperties.at("Count"));
+        count = std::get<uint64_t>(allCurrentIfaceProperties.at("Count"));
     }
-    unsigned long long deviceInstance{};
+    uint64_t deviceInstance{};
     if (allCurrentIfaceProperties.count("DeviceInstance"))
     {
-        deviceInstance = std::get<unsigned long long>(
-            allCurrentIfaceProperties.at("DeviceInstance"));
+        deviceInstance =
+            std::get<uint64_t>(allCurrentIfaceProperties.at("DeviceInstance"));
     }
     std::string inventoryObjPath{};
     if (allCurrentIfaceProperties.count("InventoryObjPath"))
